Create LTI preconfigured tool settings for using A+ LTI Tool v1.3
The LTI Tool feature was added in the A+ v1.18 release.

In Moodle, when a teacher creates an external (LTI) activity,
it is easy to set up A+ activities with the preconfigured tool settings.

In addition, change the details of the users created in the Moodle database
setup in order to match the users in the run-aplus-front container.

* email domain
* student id
* added user `root` (administrator)
* added user `newstudent` (who is not pre-created in run-aplus-front
  so that the LTI Tool launch must create the user in the A+ database)

https://github.com/apluslms/run-aplus-front