Change the Apache HTTP server's listening port to 8050 from 80
Port 8050 is needed to make local testing in Docker containers work
especially when running A+ as an LTI Tool v1.3 and Moodle as
the LTI Platform.

The port 8050 was already used in the host's web browser's address
bar in order to access the Moodle site. Now the same port is also
used in the Docker-internal network between the containers.